+++ gvm/protocols/gmpv208/entities/roles.py
@@ -1,0 +1,181 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+# pylint:  disable=redefined-builtin
+
+from typing import Any, List, Optional
+
+from gvm.errors import RequiredArgument
+from gvm.utils import add_filter, to_bool, to_comma_list
+from gvm.xml import XmlCommand
+
+
+class RolesMixin:
+    def clone_role(self, role_id: str) -> Any:
+        """Clone an existing role
+
+        Arguments:
+            role_id: UUID of an existing role to clone from
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not role_id:
+            raise RequiredArgument(
+                function=self.clone_role.__name__, argument='role_id'
+            )
+
+        cmd = XmlCommand("create_role")
+        cmd.add_element("copy", role_id)
+        return self._send_xml_command(cmd)
+
+    def create_role(
+        self,
+        name: str,
+        *,
+        comment: Optional[str] = None,
+        users: Optional[List[str]] = None,
+    ) -> Any:
+        """Create a new role
+
+        Arguments:
+            name: Name of the role
+            comment: Comment for the role
+            users: List of user names to add to the role
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+
+        if not name:
+            raise RequiredArgument(
+                function=self.create_role.__name__, argument='name'
+            )
+
+        cmd = XmlCommand("create_role")
+        cmd.add_element("name", name)
+
+        if comment:
+            cmd.add_element("comment", comment)
+
+        if users:
+            cmd.add_element("users", to_comma_list(users))
+
+        return self._send_xml_command(cmd)
+
+    def delete_role(
+        self, role_id: str, *, ultimate: Optional[bool] = False
+    ) -> Any:
+        """Deletes an existing role
+
+        Arguments:
+            role_id: UUID of the role to be deleted.
+            ultimate: Whether to remove entirely, or to the trashcan.
+        """
+        if not role_id:
+            raise RequiredArgument(
+                function=self.delete_role.__name__, argument='role_id'
+            )
+
+        cmd = XmlCommand("delete_role")
+        cmd.set_attribute("role_id", role_id)
+        cmd.set_attribute("ultimate", to_bool(ultimate))
+
+        return self._send_xml_command(cmd)
+
+    def get_roles(
+        self,
+        *,
+        filter_string: Optional[str] = None,
+        filter_id: Optional[str] = None,
+        trash: Optional[bool] = None,
+    ) -> Any:
+        """Request a list of roles
+
+        Arguments:
+            filter_string: Filter term to use for the query
+            filter_id: UUID of an existing filter to use for the query
+            trash: Whether to get the trashcan roles instead
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        cmd = XmlCommand("get_roles")
+
+        add_filter(cmd, filter_string, filter_id)
+
+        if trash is not None:
+            cmd.set_attribute("trash", to_bool(trash))
+
+        return self._send_xml_command(cmd)
+
+    def get_role(self, role_id: str) -> Any:
+        """Request a single role
+
+        Arguments:
+            role_id: UUID of an existing role
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not role_id:
+            raise RequiredArgument(
+                function=self.get_role.__name__, argument='role_id'
+            )
+
+        cmd = XmlCommand("get_roles")
+        cmd.set_attribute("role_id", role_id)
+        return self._send_xml_command(cmd)
+
+    def modify_role(
+        self,
+        role_id: str,
+        *,
+        comment: Optional[str] = None,
+        name: Optional[str] = None,
+        users: Optional[List[str]] = None,
+    ) -> Any:
+        """Modifies an existing role.
+
+        Arguments:
+            role_id: UUID of role to modify.
+            comment: Name of role.
+            name: Comment on role.
+            users: List of user names.
+
+        Returns:
+            The response. See :py:meth:`send_command` for details.
+        """
+        if not role_id:
+            raise RequiredArgument(
+                function=self.modify_role.__name__, argument='role_id argument'
+            )
+
+        cmd = XmlCommand("modify_role")
+        cmd.set_attribute("role_id", role_id)
+
+        if comment:
+            cmd.add_element("comment", comment)
+
+        if name:
+            cmd.add_element("name", name)
+
+        if users:
+            cmd.add_element("users", to_comma_list(users))
+
+        return self._send_xml_command(cmd)
